SHARED_MEM_TEST/service-a/requirements.txt
adbc-driver-manager==1.5.0
adbc-driver-postgresql==1.5.0
importlib_resources==6.5.2
numpy==2.2.4
pandas==2.2.3
psycopg2==2.9.10
pyarrow==19.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
six==1.17.0
typing_extensions==4.12.2
tzdata==2025.2
readerwriterlock==1.0.9